Consolodated API Structures
- renamed akscluster model
- fixed incorrect nodepool response mapping
- reorded constants
