+++ internal/client/akscluster/cluster_resource.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2022 VMware, Inc. All Rights Reserved.
+Copyright 2023 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
 */
