+++ internal/models/akscluster/nodepool_update.go
@@ -3,15 +3,16 @@ Copyright 2023 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
 */
 
+// Package models
+// nolint: dupl
 package models
 
-import (
-	"github.com/go-openapi/swag"
-)
+import "github.com/go-openapi/swag"
 
 // VmwareTanzuManageV1alpha1AksclusterNodepoolUpdateNodepoolRequest Request to update (overwrite) a Nodepool.
 //
 // swagger:model vmware.tanzu.manage.v1alpha1.akscluster.nodepool.UpdateNodepoolRequest
+// nolint: dupl
 type VmwareTanzuManageV1alpha1AksclusterNodepoolUpdateNodepoolRequest struct {
 
 	// Update Nodepool.
@@ -30,6 +31,36 @@ func (m *VmwareTanzuManageV1alpha1AksclusterNodepoolUpdateNodepoolRequest) Marsh
 // UnmarshalBinary interface implementation.
 func (m *VmwareTanzuManageV1alpha1AksclusterNodepoolUpdateNodepoolRequest) UnmarshalBinary(b []byte) error {
 	var res VmwareTanzuManageV1alpha1AksclusterNodepoolUpdateNodepoolRequest
+	if err := swag.ReadJSON(b, &res); err != nil {
+		return err
+	}
+
+	*m = res
+
+	return nil
+}
+
+// VmwareTanzuManageV1alpha1AksclusterNodepoolUpdateNodepoolResponse Response from updating a Nodepool.
+//
+// swagger:model vmware.tanzu.manage.v1alpha1.akscluster.nodepool.UpdateNodepoolResponse
+type VmwareTanzuManageV1alpha1AksclusterNodepoolUpdateNodepoolResponse struct {
+
+	// Nodepool updated.
+	Nodepool *VmwareTanzuManageV1alpha1AksclusterNodepoolNodepool `json:"nodepool,omitempty"`
+}
+
+// MarshalBinary interface implementation.
+func (m *VmwareTanzuManageV1alpha1AksclusterNodepoolUpdateNodepoolResponse) MarshalBinary() ([]byte, error) {
+	if m == nil {
+		return nil, nil
+	}
+
+	return swag.WriteJSON(m)
+}
+
+// UnmarshalBinary interface implementation.
+func (m *VmwareTanzuManageV1alpha1AksclusterNodepoolUpdateNodepoolResponse) UnmarshalBinary(b []byte) error {
+	var res VmwareTanzuManageV1alpha1AksclusterNodepoolUpdateNodepoolResponse
 	if err := swag.ReadJSON(b, &res); err != nil {
 		return err
 	}
