--- internal/models/akscluster/update_aks_cluster_request.go
@@ -1,40 +1,0 @@
-/*
-Copyright 2023 VMware, Inc. All Rights Reserved.
-SPDX-License-Identifier: MPL-2.0
-*/
-
-package models
-
-import (
-	"github.com/go-openapi/swag"
-)
-
-// VmwareTanzuManageV1alpha1AksclusterUpdateAksClusterRequest Request to update (overwrite) an AksCluster.
-//
-// swagger:model vmware.tanzu.manage.v1alpha1.akscluster.UpdateAksClusterRequest
-type VmwareTanzuManageV1alpha1AksclusterUpdateAksClusterRequest struct {
-
-	// Update AksCluster.
-	AksCluster *VmwareTanzuManageV1alpha1AksclusterAksCluster `json:"aksCluster,omitempty"`
-}
-
-// MarshalBinary interface implementation.
-func (m *VmwareTanzuManageV1alpha1AksclusterUpdateAksClusterRequest) MarshalBinary() ([]byte, error) {
-	if m == nil {
-		return nil, nil
-	}
-
-	return swag.WriteJSON(m)
-}
-
-// UnmarshalBinary interface implementation.
-func (m *VmwareTanzuManageV1alpha1AksclusterUpdateAksClusterRequest) UnmarshalBinary(b []byte) error {
-	var res VmwareTanzuManageV1alpha1AksclusterUpdateAksClusterRequest
-	if err := swag.ReadJSON(b, &res); err != nil {
-		return err
-	}
-
-	*m = res
-
-	return nil
-}
