+++ internal/resources/akscluster/schema.go
@@ -45,7 +45,7 @@ var ClusterSchema = map[string]*schema.Schema{
 	clusterSpecKey: ClusterSpecSchema,
 	waitKey: {
 		Type:        schema.TypeString,
-		Description: "Wait timeout duration until cluster resource reaches READY state. Accepted timeout duration values like 5s, 45m, or 3h, higher than zero",
+		Description: "Wait timeout duration until cluster resource reaches READY state. Accepted timeout duration values like 5s, 45m, or 3h, higher than zero.  The default duration is 30m",
 		Default:     "default",
 		Optional:    true,
 	},
